Crate zcash_proofs

Zcash circuits and proofs.

zcash_proofs contains the zk-SNARK circuits used by Zcash, and the APIs for creating and verifying proofs.

Modules

circuit

Implementations of the Zcash circuits and Zcash-specific gadgets.

constants

Various constants used for the Zcash proofs.

proverlocal-prover or bundled-prover

Abstractions over the proving system and parameters for ease of use.

sapling

Helpers for creating Sapling proofs.

sprout

APIs for creating and verifying Sprout proofs.

Structs

SaplingParameterPathsdownload-params

The paths to the Sapling parameter files.

ZcashParameters

Zcash Sprout and Sapling groth16 circuit parameters.

Constants

SAPLING_OUTPUT_NAME

The sapling output parameters file name.

SAPLING_SPEND_NAME

The sapling spend parameters file name.

SPROUT_NAME

The sprout parameters file name.

Functions

default_params_folderdirectories

Returns the default folder that the Zcash proving parameters are located in.

download_parametersDeprecateddownload-params

Download the Zcash Sapling parameters if needed, and store them in the default location. Always checks the sizes and hashes of the files, even if they didn’t need to be downloaded.

download_sapling_parametersdownload-params

Download the Zcash Sapling parameters if needed, and store them in the default location. Always checks the sizes and hashes of the files, even if they didn’t need to be downloaded.

download_sprout_parametersdownload-params

Download the Zcash Sprout parameters if needed, and store them in the default location. Always checks the size and hash of the file, even if it didn’t need to be downloaded.

load_parameters

Load the specified parameters, checking the sizes and hashes of the files.

parse_parameters

Parse Bls12 keys from bytes as serialized by [Parameters::write].