Crate zcash_proofs

Zcash circuits and proofs.

zcash_proofs contains the zk-SNARK circuits used by Zcash, and the APIs for creating and verifying proofs.

Modules

circuit

Implementations of the Zcash circuits and Zcash-specific gadgets.

constants

Various constants used for the Zcash proofs.

proverlocal-prover or bundled-prover

Abstractions over the proving system and parameters for ease of use.

sapling

Helpers for creating Sapling proofs.

sprout

APIs for creating and verifying Sprout proofs.

Structs

ZcashParameters

Functions

default_params_folderdirectories

Returns the default folder that the Zcash proving parameters are located in.

download_parametersdownload-params

Download the Zcash Sapling parameters, storing them in the default location.

load_parameters

parse_parameters

Parse Bls12 keys from bytes as serialized by [Parameters::write].