Module zcash_proofs::circuit::sprout

The “hybrid Sprout” circuit.

“Hybrid Sprout” refers to the implementation of the Sprout statement in bellman for groth16, instead of the original implementation using libsnark for BCTV14.

Structs

CommitmentRandomness

JoinSplit

JsInput

JsOutput

NoteValue

PayingKey

SpendingKey

UniqueRandomness

Constants

TREE_DEPTH